refactor: add package inside each module to separate gateway from usecase
