+++ application/src/main/java/br/com/jpcchaves/application/category/gateway/DeleteCategoryGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.category;
+package br.com.jpcchaves.application.category.gateway;
 
 public interface DeleteCategoryGateway {
   void delete(Long id);
