+++ application/src/main/java/br/com/jpcchaves/application/category/gateway/UpdateCategoryGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.category;
+package br.com.jpcchaves.application.category.gateway;
 
 import br.com.jpcchaves.core.domain.Category;
