+++ application/src/main/java/br/com/jpcchaves/application/todo/gateway/CreateTodoGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.todo;
+package br.com.jpcchaves.application.todo.gateway;
 
 import br.com.jpcchaves.core.domain.Todo;
