+++ application/src/main/java/br/com/jpcchaves/application/todo/gateway/DeleteTodoGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.todo;
+package br.com.jpcchaves.application.todo.gateway;
 
 public interface DeleteTodoGateway {
   void delete(Long id);
