+++ application/src/main/java/br/com/jpcchaves/application/todo/gateway/UpdateTodoGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.todo;
+package br.com.jpcchaves.application.todo.gateway;
 
 import br.com.jpcchaves.core.domain.Todo;
