+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/category/UpdateCategoryGatewayImpl.java
@@ -1,6 +1,6 @@
 package br.com.jpcchaves.infrastructure.gatewayimpl.category;
 
-import br.com.jpcchaves.application.category.UpdateCategoryGateway;
+import br.com.jpcchaves.application.category.gateway.UpdateCategoryGateway;
 import br.com.jpcchaves.core.domain.Category;
 import br.com.jpcchaves.core.exception.CategoryException;
 import br.com.jpcchaves.core.exception.enums.ExceptionDefinition;
